+++ src/main/java/std/reply/mapper/BoardMapper.java
@@ -10,11 +10,12 @@ import java.util.List;
 public interface BoardMapper {
     public List<BoardDTO> select();
 
-    public BoardDTO selectById(int id);
+    public BoardDTO selectById(Long id);
 
     public void insert(BoardDTO board);
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 
     public void update(BoardDTO board);
+
 }
