+++ src/test/java/std/reply/mapper/BoardMapperTests.java
@@ -36,14 +36,14 @@ public class BoardMapperTests {
 
     @Test
     public void testDeleteBoard() {
-        int id = 1;
+        Long id = 1L;
 
         boardMapper.deleteById(id);
     }
 
     @Test
     public void testUpdateBoard() {
-        BoardDTO boardDTO = boardMapper.selectById(2);
+        BoardDTO boardDTO = boardMapper.selectById(2L);
 
         boardDTO.setContent("update content");
         boardDTO.setTitle("update title");
